requirements.txt
accelerate
asyncio
bitsandbytes
bs4
datasets
docx
exceptions
fitz
hf_transfer
huggingface_hub
jsons
mlx-lm
numpy
openai
openpyxl
pandas
pdfplumber
peft
pymupdf
requests
selenium
sentencepiece
spacy
torch
transformers
wandb